feat(tab): ship .d.ts files with npm package
This extends the build command to copy out the .d.ts files
I also needed to add a root-level .d.ts that re-exports the .d.ts in the /react directory
