+++ packages/tab/package.json
@@ -10,15 +10,16 @@
   "module": "dist/esm/index.js",
   "scripts": {
     "build": "run-s build:js build:cjs build:css",
-    "build:cjs": "babel --root-mode upward src --out-dir dist/cjs",
+    "build:cjs": "babel --root-mode upward src --out-dir dist/cjs && cp src/react/index.d.ts dist/cjs/react/index.d.ts && cp src/index.d.ts dist/cjs/index.d.ts",
     "build:css": "build-css --useGlamor -i dist/cjs/css/index.js",
-    "build:js": "cross-env ESM=true babel --root-mode upward src --out-dir dist/esm",
+    "build:js": "cross-env ESM=true babel --root-mode upward src --out-dir dist/esm && cp src/react/index.d.ts dist/esm/react/index.d.ts && cp src/index.d.ts dist/esm/index.d.ts",
     "build:watch": "npm run build:js -- --watch",
     "prepublish": "npm run build",
     "storybook": "start-storybook -p 6006",
     "test": "jest",
     "test:watch": "npm run test -- --watchAll"
   },
+  "types": "dist/cjs/index.d.ts",
   "dependencies": {
     "@pluralsight/ps-design-system-core": "^6.0.2",
     "@pluralsight/ps-design-system-filter-react-props": "^1.1.17",
